conntrackd: basic TIPC support
Basic implementation of a TIPC channel for the conntrackd daemon
(successfully tested in NOTRACK and FTFW modes).

TIPC is a protocol that allows applications in a cluster-based
environment to communicate quickly and reliably with other
applications in the cluster.
It allows both unicast and multicast, reliable/unreliable and
datagram/stream oriented communications.

One of its main feature's of interest here is to provide sockets that
communicates in a connectionless, yet reliable manner that guarantees
delivery of every message sent over the network.

This can be useful in the context of high-available, cluster-based
firewalls where states propagation has to be both fast and reliable.

So far, the results are encouraging, though more tests have to
performed on different setups to enhance the implementation and track
any bugs.

An example config file can be found in the doc/sync/ directory of
the conntrack-tools, along with a README file providing basic
installation instructions.

Signed-off-by: Quentin Aebischer <quentin.aebischer@usherbrooke.ca>