conntrackd: remove cache feature infrastructure
This infrastructure was replaced by cache extensions (a.k.a. cache_extra)
long time ago, but there was still one single client of it.

Convert that remaining client to use the cache_extra infrastructure and
remove the obsolete cache feature infrastructure. I needed to add the
dump callback to cache_extra since it was missing.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>