conntrackd: add track mode
This new mode allows conntrackd to track entries in the kernel cache
to check for possible leak of conntrack entries. The operational is
simple:

1) Receive an event regarding a conntrack entry that resides in the
   kernel and store it in the daemon's internal cache.
2) The daemon checks if the conntrack objects that reside in the kernel
   are still there every N seconds. Being N a random amount of seconds
   between 0 and 60 seconds (the random timer ensures random timer ensures
   the timer distribution to avoid CPU peaks).
3) If the daemon does not find in the kernel some entry that resides in
   its internal cache after 5 tries, it considers that the entry has been
   leaked.

There is an example in doc/track/conntrackd-track.conf on how to setup
this working mode. By now, you cannot tweak any configuration parameter.

NetlinkEventsReliable is set on by default to turn tracking reliable,
otherwise we may hit false positives. PollSecs cannot be used with this
option.

To check for errors, consult the log files, eg.

	tail -100f /var/log/conntrackd.log

conntrackd -i shows the following output:

icmp     1 src=192.168.1.135 dst=150.214.142.14 type=8 code=0 id=23495 src=150.214.142.14 dst=192.168.1.135 type=0 code=0 id=23495 [ping in 55s, stamina 5] [active since 2s]

ping means that it will check in 55 secs again if that entry is still in the
kernel. The stamina field says how many lives this entry has, you kill one
live if you don't find the entry once in the kernel. After 5 times, the
daemon considers that entries has leaked. This helps to avoid possible false
positives due to the destroy events are still flying from the kernel to the
local internal cache yet.

This mode also monitors severe inconsistencies between the in-kernel
conntrack table counter and the number of entries in the internal cache.
In case of problem, it will spot error messages to the log file.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>